refactor: pull away nav drop menu for component
